--- packages/admin/next.config.js
@@ -1,8 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  swcMinify: true,
-};
-
-// eslint-disable-next-line no-undef
-module.exports = nextConfig;
