--- packages/admin/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@wiscro/admin",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev -p 3001",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@wiscro/common": "*"
-  },
-  "devDependencies": {
-    "@types/node": "18.15.10"
-  }
-}
