--- packages/admin/pages/_app.tsx
@@ -1,8 +1,0 @@
-import '../styles/globals.css';
-import type { AppProps } from 'next/app';
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
-
-export default MyApp;
