--- packages/common/.gitignore
@@ -1,2 +1,0 @@
-tsconfig.tsbuildinfo
-/dist
