--- packages/common/package.json
@@ -1,15 +1,0 @@
-{
-  "private": true,
-  "name": "@wiscro/common",
-  "version": "1.0.0",
-  "main": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "build:watch": "rm -fr ./dist && tsc -w"
-  },
-  "peerDependencies": {
-    "react": "*",
-    "typescript": "*"
-  }
-}
