--- packages/common/src/index.ts
@@ -1,3 +1,0 @@
-export * from './domains/Page';
-export * from './domains/shared';
-// export * from './utils'
