--- packages/common/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "sourceMap": true,
-    "rootDir": "./src",
-    "baseUrl": "./",
-    "allowJs": true,
-    "downlevelIteration": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "outDir": "./dist",
-    "declaration": true,
-  },
-  "include": ["next-env.d.ts", "src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["node_modules"]
-}
