--- packages/front/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@wiscro/front",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@wiscro/common": "*"
-  },
-  "devDependencies": {
-    "@types/node": "18.15.10"
-  }
-}
