+++ src/domains/shared/index.ts
